% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{base_dose_response}
\alias{base_dose_response}
\title{Add ggplot objects common to all dose-response plots}
\usage{
base_dose_response(
  plot,
  x_limits,
  font_base_size = 14,
  xlab = "log10[treatment] (M)",
  ylab = "\% untreated response",
  legend = TRUE,
  grid = FALSE,
  legend_title = ggplot2::waiver()
)
}
\arguments{
\item{plot}{ggplot2 plot to which to add objects}

\item{x_limits}{Limits of the x axis. Vector of form c(start, end)}

\item{font_base_size}{Font size in points. Default is 14 from theme_prism}

\item{xlab}{The label for the x axis}

\item{ylab}{The label for the y axis}

\item{legend}{Whether or not the plot should have a legend (default is TRUE)}

\item{grid}{Whether or not the plot should have a grid (default is FALSE)}

\item{legend_title}{Optional title for the legend. If not provided, will be
determined by \code{ggplot2::waiver()} and will be the name of the aesthetic
used for the legend.}
}
\value{
The plot with added objects. These are:
\itemize{
\item error bars for mean plus/minus SEM of response
\item prism theme
\item x axis scale with major breaks at log10 intervals and logistic minor breaks
\item y axis scale with zoom from 0 to max and manual 25\% breaks
\item transparent background
\item optional removed legend
\item optional major and minor grid
\item x and y labels from arguments
}
}
\description{
\code{base_dose_response()} is a helper function that adds the ggplot objects
common to doseplotr dose-response plots.
}
