% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{dir_excel_to_csv}
\alias{dir_excel_to_csv}
\title{Convert all Excel files in a directory to csv.}
\usage{
dir_excel_to_csv(dir)
}
\arguments{
\item{dir}{The directory containing Excel files to convert.}
}
\description{
\code{dir_excel_to_csv()} finds any files with .xls or .xlsx extensions in a
directory and attempts to save a CSV file of each sheet in each Excel file.
Output filenames are in the form "excelfilename_sheetname.csv". Does nothing
if no Excel files are present.
}
