% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseplotr-package.R
\docType{package}
\name{doseplotr-package}
\alias{doseplotr}
\alias{doseplotr-package}
\title{doseplotr: Import, Analyze and Visualize Dose-Response Data}
\description{
A system for analyzing and visualizing the dose-dependent effects of one or more treatments on one or more targets. You can import data for multiple combinations of treatments and targets, and 'doseplotr' can generate summary statistics (e.g. EC50) and plots for each combination of treatment and target. It can also generate more complex plots, such as a plot overlaying the effects of one treatment on all targets tested.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jackwalkerstevenson/doseplotr}
  \item Report bugs at \url{https://github.com/jackwalkerstevenson/doseplotr/issues}
}

}
\author{
\strong{Maintainer}: Jack Stevenson \email{doseplotr@gmail.com} (\href{https://orcid.org/0000-0001-7261-4958}{ORCID})

}
\keyword{internal}
