% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{EC50_log_from_point_model}
\alias{EC50_log_from_point_model}
\title{Estimate log10 EC50 from one dose-response point and 3 parameters of a model}
\usage{
EC50_log_from_point_model(model, log_dose, response)
}
\arguments{
\item{model}{A logistic model of the type returned by \code{\link[drda:drda]{drda::drda()}}}

\item{log_dose}{The dose of the single point measurement in log units}

\item{response}{The response (NOT percent inhibition) observed at the tested
dose}
}
\value{
The estimated EC50 in log10 units
}
\description{
Estimate log10 EC50 from one dose-response point and 3 parameters of a model
}
