% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{EC50_logM_from_point}
\alias{EC50_logM_from_point}
\title{Estimate the EC50 of a treatment on a target from a single measurement}
\usage{
EC50_logM_from_point(alpha, delta, eta, x, y)
}
\arguments{
\item{alpha}{The alpha parameter from a \code{drda} model.}

\item{delta}{The delta parameter from a \code{drda} model.}

\item{eta}{The eta parameter from a \code{drda} model.}

\item{x}{The dose of the single point measurement in log(molar) units}

\item{y}{The response of the single point measurement}
}
\value{
The estimated EC50 in log10(molar) units
}
\description{
Estimate the EC50 of a treatment on a target from a single measurement
}
