% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{EC50_logM_from_point_model}
\alias{EC50_logM_from_point_model}
\title{Estimate EC50 from one dose-response point and 3 parameters of a given model}
\usage{
EC50_logM_from_point_model(model, x, y)
}
\arguments{
\item{model}{A logistic model of the type returned by \code{\link[drda:drda]{drda::drda()}}}

\item{x}{The dose of the single point measurement in log(molar) units}

\item{y}{The activity (NOT percent inhibition) of the single point
measurement}
}
\value{
The estimated EC50 in log10(molar) units
}
\description{
Estimate EC50 from one dose-response point and 3 parameters of a given model
}
