% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{excel_sheet_to_csv}
\alias{excel_sheet_to_csv}
\title{Convert a sheet of an Excel file to CSV.}
\usage{
excel_sheet_to_csv(excel_path, sheet)
}
\arguments{
\item{excel_path}{Path to the Excel file. The file can be .xls or .xlsx}

\item{sheet}{The name of the sheet in the Excel file to convert to CSV}
}
\description{
\code{excel_sheet_to_csv()} reads a specific sheet of an Excel file and writes it
to a CSV file. The output filename is the name of the Excel file appended
with an underscore and the name of the sheet, i.e.
"excelfilename_sheetname.csv".
}
