% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{excel_to_csv}
\alias{excel_to_csv}
\title{Convert an Excel file to csv.}
\usage{
excel_to_csv(excel_path)
}
\arguments{
\item{excel_path}{Path to the Excel file. The file can be .xls or .xlsx}
}
\description{
Convert an Excel file to csv.
}
