% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_copy_to_dir}
\alias{file_copy_to_dir}
\title{Copy a file to a directory}
\usage{
file_copy_to_dir(path_to_file, dest_dir, timestamp = TRUE)
}
\arguments{
\item{path_to_file}{Path to the file to copy.}

\item{dest_dir}{Path to the directory into which to copy the file.}

\item{timestamp}{Whether to insert a timestamp into the output filename
(default TRUE).}
}
\description{
This method copies a file to a destination directory, optionally (default
yes) adding a timestamp before the file extension.
}
