% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_validate_reorder}
\alias{filter_validate_reorder}
\title{Filter, validate and reorder a dataframe column with desired values}
\usage{
filter_validate_reorder(df, colname, values)
}
\arguments{
\item{df}{A dataframe}

\item{colname}{The column to filter, validate and reorder}

\item{values}{The values for which to filter, check and reorder the column}
}
\value{
The dataframe with the desired column filtered, validated and
reordered
}
\description{
This function is intended for reconciling a column of a dataframe with an
externally-provided vector of desired values for the column. It does three
things:
\itemize{
\item Filter the column so it only contains the desired values
\item Validate that all of the desired values are actually present in the column.
Throws an error if any values are missing from the column.
\item Reorder the column as a factor to match the order of the desired values
}
}
\examples{
example_df <- data.frame(treatment = c("foo", "foo", "foo", "bar", "baz"),
                 target = c("apple", "orange", "banana", "apple", "orange"))
filter_validate_reorder(example_df, "target", c("banana", "orange"))
}
