% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_display_name}
\alias{get_display_name}
\title{Get the value of a name in a named vector if present and desired}
\usage{
get_display_name(name, display_names, rename)
}
\arguments{
\item{name}{The data name to be renamed}

\item{display_names}{Named vector of the form c("data name" = "display name")}

\item{rename}{Whether to seek to rename at all. This argument is intended for
use with a global parameter that determines whether display names will be
applied across all plots.}
}
\value{
The name to use for the group in the plot. Will be either the
original name or a new display name depending on arguments.
}
\description{
This function is intended for setting the "display name" of a group in a plot
from its "data name", its name in the underlying dataframe.
}
\examples{
example_names <- c("foo" = "Display Name for Foo",
                   "bar" = "Display Name for Bar")
get_display_name("foo", example_names, TRUE)
get_display_name("foo", example_names, FALSE)
get_display_name("baz", example_names, TRUE)
}
