% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda.R
\name{get_drda}
\alias{get_drda}
\title{Get a drda dose-response model for the activity of a treatment on a target}
\usage{
get_drda(data, activity_colname = "activity", ...)
}
\arguments{
\item{data}{A dataframe containing the following columns:
\itemize{
\item "activity" (default) or other activity column name (see \code{activity_colname})
\item conc_logM: concentration of treatment in log(molar) units
}}

\item{activity_colname}{Name of the column containing activity of treatment.
Default is "activity". Typical alternatives might include e.g.
"percent_inhibition", "viability", "growth".}

\item{...}{Additional parameters to pass on to \code{\link[drda:drda]{drda::drda()}}}
}
\value{
A dose-response model object of class \code{drda}.
}
\description{
\code{get_drda()} runs \code{\link[drda:drda]{drda::drda()}} to generate a 4-parameter logistic
dose-response model for the effect of a given treatment on a target.
}
