% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drda.R
\name{get_drda_fixed}
\alias{get_drda_fixed}
\title{Fit a drda 4-parameter logistic model with one or more parameters fixed}
\usage{
get_drda_fixed(
  data,
  alpha = NA,
  delta = NA,
  eta = NA,
  phi = NA,
  activity_col = "activity"
)
}
\arguments{
\item{data}{A dataframe containing dose-response data for only one treatment
on one target. If needed, filter data in advance with \code{\link[=filter_trt_tgt]{filter_trt_tgt()}}.
The dataframe should contain the following columns:
\itemize{
\item "activity" (default) or other activity column name (see \code{activity_col})
\item conc_logM: concentration of treatment in log(molar) units
}}

\item{alpha}{Optional: fixed value of asymptote level at 0 dose. See
\code{\link[drda:drda]{drda::drda()}}.}

\item{delta}{Optional: fixed value of height of the curve (difference betwene
minimum and maximum concentration asymptotes). Signed (can be positive or
negative). See \code{\link[drda:drda]{drda::drda()}}.}

\item{eta}{Optional: fixed value of growth rate or Hill slope. Not signed
(always positive). See \code{\link[drda:drda]{drda::drda()}}.}

\item{phi}{Optional: fixed value of mid-value or EC50. See \code{\link[drda:drda]{drda::drda()}}.}

\item{activity_col}{Name of the column containing activity of treatment.
Default is "activity". Typical alternatives might include e.g.
"percent_inhibition", "viability", "growth".}
}
\value{
A dose-response model object of class \code{drda}.
}
\description{
\code{get_drda_fixed()} runs \code{\link[=get_drda_helper]{get_drda_helper()}} to generate a 4-parameter
logistic dose-response model for the effect of a given treatment on a target
with one or more of the parameters fixed at a given value. The four
parameters are available as optional arguments, named as in \code{\link[drda:drda]{drda::drda()}}.
}
