% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drda.R
\name{get_drda_helper}
\alias{get_drda_helper}
\title{Get a drda dose-response model without bounds on parameters}
\usage{
get_drda_helper(data, response_col = "response", ...)
}
\arguments{
\item{data}{A dataframe containing dose-response data for only one treatment
on one target. If needed, filter data in advance with \code{\link[=filter_trt_tgt]{filter_trt_tgt()}}.
The dataframe should contain the following columns:
\itemize{
\item "response" (default) or other response column name (see \code{response_col})
\item log_dose: dose of treatment in log10 units
}}

\item{response_col}{Name of the column containing response values.
Default is "response". Typical alternatives might include e.g.
"percent_inhibition", "viability", "growth", "activity".}

\item{...}{Additional arguments to be passed to \code{\link[drda:drda]{drda::drda()}}, e.g. bounds.}
}
\value{
A dose-response model object of class \code{drda}.
}
\description{
\code{get_drda_helper()} runs \code{\link[drda:drda]{drda::drda()}} to generate a 4-parameter logistic
dose-response model for the effect of a given treatment on a target. It
places no bounds on the parameters by default but allows arguments to be
passed to \code{\link[drda:drda]{drda::drda()}}. It is intended for internal use in generating
bounded models.
}
