% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda.R
\name{get_drda_unbounded}
\alias{get_drda_unbounded}
\title{Get a drda dose-response model without bounds on parameters}
\usage{
get_drda_unbounded(data, activity_col = "activity")
}
\arguments{
\item{data}{A dataframe containing the following columns:
\itemize{
\item "activity" (default) or other activity column name (see \code{activity_col})
\item conc_logM: concentration of treatment in log(molar) units
}}

\item{activity_col}{Name of the column containing activity of treatment.
Default is "activity". Typical alternatives might include e.g.
"percent_inhibition", "viability", "growth".}
}
\value{
A dose-response model object of class \code{drda}.
}
\description{
\code{get_drda_unbounded()} runs \code{\link[drda:drda]{drda::drda()}} to generate a 4-parameter logistic
dose-response model for the effect of a given treatment on a target. It
places no bounds on the parameters. It is intended for internal use in
generating bounded models.
}
