% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drm.R
\name{get_drm}
\alias{get_drm}
\title{Get a drc dose-response model for the activity of a treatment on a target}
\usage{
get_drm(data, activity_col = "activity", ...)
}
\arguments{
\item{data}{A dataframe containing the following columns:
\itemize{
\item "activity" (default) or other activity column name (see \code{activity_col})
\item conc_logM: concentration of treatment in log(molar) units
}}

\item{activity_col}{Name of the column containing activity of treatment.
Default is "activity". Typical alternatives might include e.g.
"percent_inhibition", "viability", "growth".}

\item{...}{Additional parameters to pass on to \code{\link[drc:drm]{drc::drm()}}}
}
\value{
A dose-response model object of class \code{drc}.
}
\description{
\code{get_drm()} runs \code{\link[drc:drm]{drc::drm()}} to generate a 4-parameter logistic
dose-response model for the effect of a given treatment on a target.
}
