% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{get_EC_log}
\alias{get_EC_log}
\title{Get effective log10 dose for a percent of maximum response from a drda model}
\usage{
get_EC_log(model, level = 50)
}
\arguments{
\item{model}{A logistic model of the type returned by \code{\link[drda:drda]{drda::drda()}}}

\item{level}{The desired response in percent. Percent of the way from the
minimum-dose response to the maximum-dose response, regardless of whether
the response is positive or negative. Default is 50, giving the classic
EC50.}
}
\value{
The effective dose in log10 units.
}
\description{
This function returns the dose for which a dose-response model
gives a desired fraction of the maximum response. For instance, the EC50 is
the dose that gives 50\% of the maximum response, regardless of the direction
or magnitude of the response. This is not to be confused with the IC50, which
is the dose that gives a response of exactly 50 and is given by
\code{get_IC_log()}.
}
