% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{get_EC_logM}
\alias{get_EC_logM}
\title{Get effective logmolar concentration for a response level from a drda model}
\usage{
get_EC_logM(model, level = 50)
}
\arguments{
\item{model}{A logistic model of the type returned by \code{\link[drda:drda]{drda::drda()}}}

\item{level}{The desired response level in percent. Percent of the way from
the minimum-dose response to the maximum-dose response, regardless of
whether the response is positive or negative. Default is 50, giving the
classic EC50.}
}
\value{
The effective concentration in log10(molar) units.
}
\description{
Get effective logmolar concentration for a response level from a drda model
}
