% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drm.R
\name{get_EC_logM_drm}
\alias{get_EC_logM_drm}
\title{Title Get an EC value with drc for the activity of a treatment on a target}
\usage{
get_EC_logM_drm(
  data,
  trt,
  tgt,
  response_level = 50,
  trt_col = "treatment",
  tgt_col = "target",
  activity_col = "activity"
)
}
\arguments{
\item{data}{A dataframe containing the following columns:
\itemize{
\item "treatment" (default) or other treatment column name (see \code{trt_col})
\item "target" (default) or other target column name (see \code{tgt_col})
}}

\item{trt}{Name of the treatment to use in the model}

\item{tgt}{Name of the target to use in the model}

\item{response_level}{The response level in percent for which to estimate the
effective concentration. This parameter is used as the \code{respLev} parameter
of \code{\link[drc:ED.drc]{drc::ED()}}. Should be between 0 and 100.

Be careful: this parameter can be confusing when the response to treatment
is a decrease in activity. This is because the response level represents
the percent of the way from whichever asymptote is a lower activity level
to whichever is higher rather than from the zero-dose asymptote to the
infinite-dose asymptote. For example, if activity starts at 100 at 0 dose
and goes down to 0 at infinite dose, then \code{get_ED_logM(response_level=75)}
will return the concentration that gives an activity of 75, even though you
might naively expect that it would return the concentration that gives an
activity of 25 (75\% of the way to the infinite-dose effect). This is
debatably a bug in \code{\link[drc:ED.drc]{drc::ED()}}.

Be careful: this parameter can also be confusing for data not normalized to
100. For example, if activity e.g. starts at 200 at 0 dose and increases to
500 at infinite dose, then \code{get_ED_logM_drm(response_level=50)} will return
the concentration that gives an activity of 350 (50\% of the way from 200 to
500).}

\item{trt_col}{Name of the column containing treatments. Default is
"treatment". Typical alternatives might include e.g. "drug", "additive",
"supplement".}

\item{tgt_col}{Name of the column containing targets. Default is "target".
Typical alternatives might include e.g. "mutant", "cell_line", "species".}

\item{activity_col}{Name of the column containing activity of treatment.
Default is "activity". Typical alternatives might include e.g.
"percent_inhibition", "viability", "growth".}
}
\value{
The effective concentration value at the desired response level in
log10(molar) units.
}
\description{
\code{get_EC_logM_drm()} runs \code{\link[=filter_trt_tgt]{filter_trt_tgt()}} and \code{\link[=get_drm]{get_drm()}} to get a
dose-response model of the effect of conc_logM on activity for the specified
treatment and target, then extracts the effective concentration value of the
model at the specified threshold using \code{\link[drc:ED.drc]{drc::ED()}}.
}
