% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{get_EC_nM}
\alias{get_EC_nM}
\title{Get effective nanomolar concentration for a response level from a drda model}
\usage{
get_EC_nM(model, level = 50)
}
\arguments{
\item{model}{A model of type \code{drda}.}

\item{level}{The desired response level in percent. Percent of the way from
the minimum-dose response to the maximum-dose response, regardless of
whether the response is positive or negative. Default is 50, giving the
classic EC50.}
}
\value{
The effective concentration in nanomolar units.
}
\description{
Get effective nanomolar concentration for a response level from a drda model
}
