% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{get_high_dose_asymptote}
\alias{get_high_dose_asymptote}
\title{Get the high-dose asymptote from a \code{drda} model}
\usage{
get_high_dose_asymptote(model)
}
\arguments{
\item{model}{A logistic model of the type returned by \code{\link[drda:drda]{drda::drda()}}}
}
\value{
The level of the high-dose asymptote. The value of the model when
dose -> Inf.
}
\description{
Get the high-dose asymptote from a \code{drda} model
}
