% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{get_hill_slope}
\alias{get_hill_slope}
\title{Get the Hill slope from a \code{drda} model}
\usage{
get_hill_slope(model)
}
\arguments{
\item{model}{A logistic model of the type returned by \code{\link[drda:drda]{drda::drda()}}}
}
\value{
The Hill slope (growth rate) of the model. Unsigned (always
positive).
}
\description{
Get the Hill slope from a \code{drda} model
}
