% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{get_IC_logM}
\alias{get_IC_logM}
\title{Get effective log10 dose for an exact response level from a drda model}
\usage{
get_IC_logM(model, level = 50)
}
\arguments{
\item{model}{A logistic model of the type returned by \code{\link[drda:drda]{drda::drda()}}}

\item{level}{The desired exact response level.}
}
\value{
The effective dose in log10 units.
}
\description{
This function returns the concentration for which a dose-response model
gives an exact desired response level. For instance, the IC50 is the dose
that gives a response level of exactly 50. This is not to be confused with
the EC50, which is the dose that gives 50\% of the maximum response and is
given by \code{get_EC_logM()}.
}
