% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_if}
\alias{get_if}
\title{Get a variable if a boolean is TRUE, or else a backup option}
\usage{
get_if(var, bool, otherwise = NULL)
}
\arguments{
\item{var}{Variable to return if \code{bool} is TRUE}

\item{bool}{Whether to return \code{var} or \code{otherwise}}

\item{otherwise}{What to return if \code{bool} is FALSE. Default is NULL.}
}
\value{
\code{var}, if \code{bool} is TRUE, else \code{otherwise}
}
\description{
Get a variable if a boolean is TRUE, or else a backup option
}
