% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{get_low_dose_asymptote}
\alias{get_low_dose_asymptote}
\title{Get the low-dose asymptote from a \code{drda} model}
\usage{
get_low_dose_asymptote(model)
}
\arguments{
\item{model}{A logistic model of the type returned by \code{\link[drda:drda]{drda::drda()}}}
}
\value{
The level of the low-dose asymptote. The value of the model when dose
-> -Inf.
}
\description{
Get the low-dose asymptote from a \code{drda} model
}
