% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{get_predictions}
\alias{get_predictions}
\title{Get a dataframe of predictions from desired doses and a dataframe of models}
\usage{
get_predictions(models_df, response_col = "mean_response")
}
\arguments{
\item{models_df}{Dataframe containing models and information about them. Has columns:
\itemize{
\item "treatment": treatment for which the model was fit
\item "target": target for which the model was fit
\item "model": dose-response model fit for the specified treatment and target
}}

\item{response_col}{Name for the column of predicted responses}
}
\value{
A dataframe containing predicted values. Has columns:
\itemize{
\item treatment
\item target
\item log_dose
\item response (or other column name specified in response_col)
}
}
\description{
Get a dataframe of predictions from desired doses and a dataframe of models
}
