% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_timestamp}
\alias{get_timestamp}
\title{Get a string representation of current system time}
\usage{
get_timestamp()
}
\value{
A string representing current system time in YYYY-MM-DDTHHMMSS
format, which is a filesystem-friendly adaptation of ISO 8601 without
colons. For example, 8:17 pm and 40 seconds on July 20, 1969 is represented
as 1969-06-20T201740.
}
\description{
Get a string representation of current system time
}
\examples{
get_timestamp()
}
