% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_kinomescan}
\alias{import_kinomescan}
\title{Import and preprocess a KINOMEscan raw data report}
\usage{
import_kinomescan(input_filename)
}
\arguments{
\item{input_filename}{Path to file to import. \code{import_kinomescan()} expects
the type of file that Eurofins KINOMEscan offers as a csv download as of
2023, with the column names unedited. Optionally, a "treatment" column can
be added containing names to replace the names in the "Compound Name"
column. Otherwise that column is renamed to "treatment". It expects the
following column names:
\itemize{
\item "Compound.Name" (or "treatment")
\item "DiscoveRx Gene Symbol" (or "target")
\item "Compound Concentration (nM)" (or "conc_nM")
\item "Signal" (or "response")
}}
}
\value{
The imported data as a tibble. Contains the following column names:
\itemize{
\item "treatment"
\item "target"
\item "response"
\item "dose_nM"
\item "log_dose
}
}
\description{
Import and preprocess a KINOMEscan raw data report
}
