% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_plater_CSVs}
\alias{import_plater_CSVs}
\title{Import an entire directory of \code{plater}-formatted CSVs}
\usage{
import_plater_CSVs(dir)
}
\arguments{
\item{dir}{The directory from which to import all CSV files.}
}
\value{
A data frame containing the combined data from all the files.
}
\description{
\code{import_plater_CSVs()} attempts to use \code{plater::read_plates()} to import
every CSV file in the given directory. Will only work if every CSV file is in
proper \code{plater} format. Does not process the data.
}
