% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_plates}
\alias{import_plates}
\title{Import and preprocess an entire directory of plate-based dose-response data}
\usage{
import_plates(dir)
}
\arguments{
\item{dir}{The directory from which to import all CSV files.}
}
\value{
A data frame containing combined, processed data from all the files
}
\description{
\code{import_plates()} reads an entire directory of \code{plater}-formatted Excel or
CSV files representing plate-based dose-response data. If any Excel files are
present they are converted to CSVs, one per sheet, then all CSV files are
read.
}
\details{
Data is preprocessed (renamed and normalized) for later analysis with
\code{preprocess_plate_data()}.
}
