% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_selectscreen}
\alias{import_selectscreen}
\title{Import and preprocess a csv file of the type generated by SelectScreen.}
\usage{
import_selectscreen(input_filename)
}
\arguments{
\item{input_filename}{Path to file to import. \code{import_selectscreen()} expects
the type of file that Thermo SelectScreen offers as a csv download as of
2023, with the column names unedited. It expects the following columns:
\itemize{
\item Compound.Name (or "treatment")
\item Kinase (or "target")
\item \% Inhibition 1
\item \% Inhibition 2
}}
}
\value{
The imported table as a tibble.
}
\description{
\code{import_selectscreen()} expects to receive an unedited SelectScreen results
file. It uses \code{\link[readr:read_delim]{readr::read_csv()}}'s \code{name_repair="universal"} option to make
the column names unique and syntactic, then renames columns commonly used in
dose-response analysis to simpler, more readable names:
}
\details{
\itemize{
\item "Compound.Name" -> "treatment"
\item "Kinase" -> "target"
}

If "treatment" or "target" columns are already present they will not be
overwritten.

The two untidy "\% Inhibition" columns are pivoted to one tidy
"pct_inhibition" column, the "Compound.Conc" column (which SelectScreen
indicates is in nM units) is converted to log10(molar), and a "response"
column is generated by subtracting the percent inhibition from 100.
}
