% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logM_to_nM}
\alias{logM_to_nM}
\title{Convert concentration from log(molar) to nanomolar units}
\usage{
logM_to_nM(conc_logmolar)
}
\arguments{
\item{conc_logmolar}{Concentration in log(molar) units}
}
\value{
A number: the concentration in nM
}
\description{
Convert concentration from log(molar) to nanomolar units
}
\examples{
logmolar_to_nM(-9)
}
