% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logmolar_to_nM}
\alias{logmolar_to_nM}
\title{Convert concentration from log(molar) to nanomolar units}
\usage{
logmolar_to_nM(conc_logmolar)
}
\arguments{
\item{conc_logmolar}{Concentration in log(molar) units}
}
\value{
Concentration in nM
}
\description{
Convert concentration from log(molar) to nanomolar units
}
\examples{
logmolar_to_nM(-9)
}
