% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longest}
\alias{longest}
\title{Get the length of the longest string in a vector of strings}
\usage{
longest(strings)
}
\arguments{
\item{strings}{A vector of strings}
}
\value{
Length of the longest string in the vector
}
\description{
Get the length of the longest string in a vector of strings
}
