% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{make_log_conc}
\alias{make_log_conc}
\title{Convert conc_uM or conc_nM columns to log(molar)}
\usage{
make_log_conc(df)
}
\arguments{
\item{df}{A dataframe containing either a column called "conc_nM" or a column
called "conc_uM", representing concentration in nanomolar or micromolar
units respectively.}
}
\value{
The same dataframe with an additional column "conc_logM" representing
concentration in log(molar) units.
}
\description{
Convert conc_uM or conc_nM columns to log(molar)
}
\examples{
df <- data.frame("treatment" = c("foo", "bar", "baz"),
                 "conc_nM" = c(1, 10, 100))
make_log_conc(df)
}
