% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetic_utils.R
\name{minor_breaks}
\alias{minor_breaks}
\title{Generate minor breaks for an axis in non-log-transformed units}
\usage{
minor_breaks(min_val, max_val, log_units = FALSE)
}
\arguments{
\item{min_val}{The minimum value to plot, or the log floor of same (0.1 and
0.3 will both be converted to 0.1).}

\item{max_val}{The maximum value to plot, or the log ceiling of same (500 and
1000 will both be converted to 1000).}

\item{log_units}{Whether the units of the minimum and maximum values are in
log-transformed units. Default is FALSE.}
}
\value{
A vector of minor axis breaks in raw units (not log-transformed).
}
\description{
This function generates logistic minor axis breaks to use on an axis that is
in original, non-log-transformed units, such that the data being plotted is
e.g. "1000" rather than "3".
}
\examples{
minor_breaks(1, 1000)
minor_breaks(3, 857)
minor_breaks(-12, -4, log = TRUE)
minor_breaks(-12.5, -3.8, log = TRUE)
}
