% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nM_to_logmolar}
\alias{nM_to_logmolar}
\title{Convert concentration from nanomolar to log(molar) units}
\usage{
nM_to_logmolar(conc_nM)
}
\arguments{
\item{conc_nM}{Concentration in nanomolar units}
}
\value{
Concentration in log(molar)
}
\description{
Convert concentration from nanomolar to log(molar) units
}
\examples{
nM_to_logmolar(1)
}
