% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{normalize_dose_response}
\alias{normalize_dose_response}
\title{Normalize dose-response data to 0-concentration conditions}
\usage{
normalize_dose_response(df, .col_to_norm = "response")
}
\arguments{
\item{df}{A dataframe containing dose-response data. Should contain the
following columns:
\itemize{
\item "treatment", the condition being dosed
\item "target", the target of the treatment, e.g. a cell line
\item "conc_logM", the dose of the treatment in log10(molar) units. Should
include control data for 0 concentration (-Inf in log(molar) units) for
each treatment condition.
\item A column of response data, the name of which is indicated by
\code{.col_to_norm}.
}}

\item{.col_to_norm}{The name of the column to normalize within each treatment
condition. Default is "response".}
}
\value{
The input dataframe with the addition of a new column of normalized
data. The name of the new column is the name of \code{col_to_norm} appended with
an underscore and "norm", e.g. "response_norm".
}
\description{
\code{normalize_dose_response()} independently normalizes each treatment/target
combination in a dataset of dose-response data to its own set of control
values. It is intended to be used for situations such as plate-based assays
where each plate contains replicate dose series that include control wells,
so that is it appropriate to normalize the data from each treatment to the
mean of the untreated wells of the same condition on the same plate.
}
\details{
Data from each treatment/target combination is normalized to the average of
the control data points for that treatment/target combination, which are
indicated by a concentration of 0. This function assumes that each treatment
has its own control data and does not support sharing control data points
between treatments. Throws an error if any condition has no controls.
}
