% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{param_bounds}
\alias{param_bounds}
\title{Replace NA values in a vector with Inf or -Inf}
\usage{
param_bounds(params, lower = FALSE)
}
\arguments{
\item{params}{A vector of values (generally representing the four parameters
of the logistic function in \code{\link[drda:drda]{drda::drda()}}).}

\item{lower}{Whether to generate Inf or -Inf values for undefined parameters.
Positive is default.}
}
\value{
A vector of the same length as the original vector. NA values in the
input vector are replaced with Inf or -Inf.
}
\description{
This method is intended to prepare vectors for use by \code{\link[=get_drda_fixed]{get_drda_fixed()}} as
the arguments for \code{\link[drda:drda]{drda::drda()}}'s \code{lower_bound} and \code{upper_bound}
parameters.
}
