% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{predict_helper}
\alias{predict_helper}
\title{Get a dataframe of the predictions of a single dose-response model}
\usage{
predict_helper(model, dose_seq, trt, tgt, response_col)
}
\arguments{
\item{model}{The model to use to predict responses}

\item{dose_seq}{The doses for which to predict responses}

\item{trt}{Name of the treatment used to fit the model}

\item{tgt}{Name of the target used to fit the model}

\item{response_col}{Name of the column containing predicted responses}
}
\value{
A dataframe containing the predictions of the model. Has columns:
\itemize{
\item treatment
\item target
\item log_dose
\item response (or other column name specified in response_col)
If no model is passed, returns null.
}
}
\description{
\code{predict_helper()} is a helper function for \code{get_predictions()}. Its logic only depends on \code{model} and \code{dose_seq}, but it also takes \code{trt}, \code{tgt} and \code{response_col} just so that the returned dataframe will have proper column names.
}
