% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_plate_data}
\alias{preprocess_plate_data}
\title{Preprocess plate-based dose-response data}
\usage{
preprocess_plate_data(df)
}
\arguments{
\item{df}{A dataframe, such as the output of \code{import_plater_CSVs()},
containing raw plate data. Should contain the following columns:
\itemize{
\item "treatment"
\item "target"
\item "dose_nM" or "dose_uM"
\item "response"
}}
}
\value{
A dataframe containing processed data.
}
\description{
Preprocess plate-based dose-response data
}
