% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot as an image with predictable plot dimensions}
\usage{
save_plot(
  plot,
  filename,
  legend = TRUE,
  legend_len = NULL,
  font_base_size = 14,
  nrow = 1,
  ncol = 1,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{plot}{A ggplot2 object}

\item{filename}{Path to the file to be saved}

\item{legend}{Whether or not the plot should have a legend (default is TRUE)}

\item{legend_len}{Length of the longest piece of text in the legend}

\item{font_base_size}{Size (point units) of font in plots. 14 is theme_prism
default}

\item{nrow}{Number of rows of facets in the plot}

\item{ncol}{Number of columns of facets in the plot}

\item{width}{Width of the image to be saved}

\item{height}{Height of the image to be saved}
}
\description{
\code{save_plot()} takes a ggplot2 object and saves it as an image.
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} allows the dimensions of the output image to be specified
but does not offer logic to make the actual plot portion of the image a
constant size. \code{save_plot()} tries to set the exported image dimensions such
that the plot portion of the image is a similar size across exports, even
when plots have legends of different sizes, using parameters empirically
optimized for dose-response curve plots.
}
