% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetic_utils.R
\name{shape_scale}
\alias{shape_scale}
\title{Get a manually-curated vector of ggplot shape IDs}
\usage{
shape_scale()
}
\value{
A vector of numbers intended for use in
\code{ggplot2::scale_shape_manual()}
}
\description{
Get a manually-curated vector of ggplot shape IDs
}
\examples{
shape_scale()
}
