% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetic_utils.R
\name{shape_scale_default}
\alias{shape_scale_default}
\title{Get a manually-curated general-purpose shape scale for ggplot}
\usage{
shape_scale_default()
}
\value{
A vector of numbers corresponding to shapes intended for use in
\code{ggplot2::scale_shape_manual()} for general-purpose shape scales
}
\description{
Get a manually-curated general-purpose shape scale for ggplot
}
\examples{
shape_scale_default()
}
