% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_summarize.R
\name{summarize_models}
\alias{summarize_models}
\title{Summarize model parameters for all combinations of treatments and targets}
\usage{
summarize_models(data, activity_col = "activity")
}
\arguments{
\item{data}{A dataframe containing dose-response data for one or more
combinations of treatments and targets. The dataframe should contain the
following columns:
\itemize{
\item "activity" (default) or other activity column name (see \code{activity_col})
\item conc_logM: concentration of treatment in log(molar) units
}}

\item{activity_col}{Name of the column containing activity of treatment.
Default is "activity". Typical alternatives might include e.g.
"percent_inhibition", "viability", "growth".}
}
\value{
A summary dataframe containing one row for each combination of
treatment and target in \code{data} and containing the following columns:
\itemize{
\item EC50_nM
\item EC75_nM: the dose at which 75\% of the maximum response is achieved, whether
the response is positive or negative.
\item hill_slope: the \code{eta} parameter from \code{\link[drda:drda]{drda::drda()}}. The rate of growth of
the model, unsigned (always positive).
}
}
\description{
Summarize model parameters for all combinations of treatments and targets
}
