% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{summarize_response}
\alias{summarize_response}
\title{Summarize replicate data for plotting}
\usage{
summarize_response(df, response_col = "response")
}
\arguments{
\item{df}{Dataframe containing the data to plot}

\item{response_col}{Name of the column containing response data to summarize}
}
\value{
A dataframe containing the data summarized by \code{dplyr::summarize()}.
Columns are:
\itemize{
\item "sem": the standard error of the mean of the response values
\item "mean_response": the mean of the response values
\item "w": the width values for error bars.
This is a workaround for \code{ggplot2::ggplot()} behavior where the width of
error bars depends on the size of the group.
}
}
\description{
Summarize replicate data for plotting
}
