% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{base_dose_response}
\alias{base_dose_response}
\title{Add ggplot objects common to all dose-response plots}
\usage{
base_dose_response(plot, x_limits, font_base_size = 14, legend = TRUE)
}
\arguments{
\item{plot}{ggplot2 plot to which to add objects}

\item{x_limits}{Limits of the x axis. Vector of form c(start, end)}

\item{font_base_size}{Font size in points. Default is 14 from theme_prism}

\item{legend}{Whether or not the plot should have a legend (default is TRUE)}
}
\value{
The plot with added objects
}
\description{
Add ggplot objects common to all dose-response plots
}
