% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda_utils.R
\name{EC50_logM_from_point_params}
\alias{EC50_logM_from_point_params}
\title{Estimate EC50 from one dose-response point and given model parameters}
\usage{
EC50_logM_from_point_params(alpha, delta, eta, log_dose, activity)
}
\arguments{
\item{alpha}{The alpha parameter from a \code{drda} model.}

\item{delta}{The delta parameter from a \code{drda} model.}

\item{eta}{The eta parameter from a \code{drda} model.}

\item{log_dose}{The dose of the single point measurement in log units}

\item{activity}{The activity (NOT percent inhibition) observed at the tested
dose}
}
\value{
The estimated EC50 in log10 units
}
\description{
Estimate EC50 from one dose-response point and given model parameters
}
