% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drda.R
\name{get_drda}
\alias{get_drda}
\title{Get a dose-response model from data with flexible bounds on parameters}
\usage{
get_drda(data, response_col = "response")
}
\arguments{
\item{data}{A dataframe containing dose-response data for only one treatment
on one target. If needed, filter data in advance with \code{\link[=filter_trt_tgt]{filter_trt_tgt()}}.
The dataframe should contain the following columns:
\itemize{
\item "response" (default) or other response column name (see \code{response_col})
\item log_dose: dose of treatment in log10 units
}}

\item{response_col}{Name of the column containing response values.
Default is "response". Typical alternatives might include e.g.
"percent_inhibition", "viability", "growth", "activity".}
}
\value{
A dose-response model object of class \code{drda}.
}
\description{
\code{get_drda()} runs \code{\link[=get_drda_helper]{get_drda_helper()}} to generate a 4-parameter logistic
dose-response model for the effect of a given treatment on a target. It
attempts to bound the parameters of the model to avoid physically unrealistic
fits but allows some variation in 0-dose asymptote value.
}
\details{
If the height of the model is negative (if response decreases at higher
dose), the zero-dose asymptote is bounded in (80, 120), and the height is
bounded at the negative height of the zero-dose asymptote to avoid producing
models with infinite-dose asymptotes far below 0. Note that after the model
is refit with this bound, the height of the zero-dose asymptote may decrease
slightly, so the infinite-dose asymptote may still end up slightly below 0.

If the height of the model is positive (if response increases at higher
dose), the zero-dose asymptote is bounded in (-20, 20), and the height is
bounded at 100 minus the height of the 0-dose asymptote to avoid producing
models with infinite-dose asymptotes far above 100. This is only valid for
response units like percent inhibition that should not realistically exceed
100. Note that after the model is refit with this bound, the height of the
0-dose asymptote may increase slightly, so the maximum-dose asymptote may
still end up slightly above 100.
}
