% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drda.R
\name{get_drda}
\alias{get_drda}
\title{Get a drda dose-response model for the activity of a treatment on a target}
\usage{
get_drda(data, activity_col = "activity")
}
\arguments{
\item{data}{A dataframe containing dose-response data for only one treatment
on one target. If needed, filter data in advance with \code{\link[=filter_trt_tgt]{filter_trt_tgt()}}.
The dataframe should contain the following columns:
\itemize{
\item "activity" (default) or other activity column name (see \code{activity_col})
\item conc_logM: concentration of treatment in log(molar) units
}}

\item{activity_col}{Name of the column containing activity of treatment.
Default is "activity". Typical alternatives might include e.g.
"percent_inhibition", "viability", "growth".}
}
\value{
A dose-response model object of class \code{drda}.
}
\description{
\code{get_drda()} runs \code{\link[=get_drda_helper]{get_drda_helper()}} to generate a 4-parameter logistic
dose-response model for the effect of a given treatment on a target. It
attempts to bound the parameters of the model to avoid physically unrealistic
fits.
}
\details{
If the height of the model is negative (if activity decreases at higher
dose), the height is bounded at the height of the unbounded 0-concentration
asymptote to avoid producing models with infinite-concentration asymptotes
far below 0. Note that after the model is refit with this bound, the height
of the 0-concentration asymptote may decrease slightly, so the
maximum-concentration asymptote may still end up slightly below 0.

If the height of the model is positive (if activity increases at higher
dose), the height is bounded at 100 minus the height of the unbounded
0-concentration asymptote to avoid producing models with
infinite-concentration asymptotes far above 100. This is only valid for
activity units like percent inhibition that should not realistically exceed
100. Note that after the model is refit with this bound, the height of the
0-concentration asymptote may increase slightly, so the maximum-concentration
asymptote may still end up slightly above 100.
}
