% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drm.R
\name{get_drm}
\alias{get_drm}
\title{Get a dose-response model object for the activity of a treatment on a target}
\usage{
get_drm(data, trt, tgt)
}
\arguments{
\item{data}{A dataframe containing the following columns:
\itemize{
\item treatment
\item target
\item activity: the activity of a treatment on a target, e.g. percent inhibition
\item conc_logM: concentration of treatment in log(molar) units
}}

\item{trt}{Name of the treatment to use in the model}

\item{tgt}{Name of the target to use in the model}
}
\value{
A dose-response model object of class \code{drc}.
}
\description{
\code{get_drm()} runs \code{\link[drc:drm]{drc::drm()}} to generate a 4-parameter logistic
dose-response model for the effect of a given treatment on a target. It takes
a dataframe that can include data for more than just the desired combination
of treatment and target. The dataframe is filtered to contain only data for
the desired combination before the model is fit.
}
