% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drm.R
\name{get_drm}
\alias{get_drm}
\title{Get a dose-response model object for the activity of a treatment on a target}
\usage{
get_drm(
  data,
  trt,
  tgt,
  trt_colname = "treatment",
  tgt_colname = "target",
  activity_colname = "activity"
)
}
\arguments{
\item{data}{A dataframe containing the following columns:
\itemize{
\item "treatment" (default) or other treatment column name (see \code{trt_colname})
\item "target" (default) or other target column name (see \code{tgt_colname})
\item "activity" (default) or other activity column name (see \code{activity_colname})
\item conc_logM: concentration of treatment in log(molar) units
}}

\item{trt}{Name of the treatment to use in the model}

\item{tgt}{Name of the target to use in the model}

\item{trt_colname}{Name of the column containing treatments.
Default is "treatment". Typical alternatives might include e.g. "drug",
"additive", "supplement".}

\item{tgt_colname}{Name of the column containing targets.
Default is "target". Typical alternatives might include e.g. "mutant",
"cell_line", "species".}

\item{activity_colname}{Name of the column containing activity of treatment.
Default is "activity". Typical alternatives might include e.g.
"percent_inhibition", "viability", "growth".}
}
\value{
A dose-response model object of class \code{drc}.
}
\description{
\code{get_drm()} runs \code{\link[drc:drm]{drc::drm()}} to generate a 4-parameter logistic
dose-response model for the effect of a given treatment on a target. It takes
a dataframe that can include data for more than just the desired combination
of treatment and target. The dataframe is filtered to contain only data for
the desired combination before the model is fit.
}
