% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_target}
\alias{plot_target}
\title{Plot dose-response effects of one target on all treatments}
\usage{
plot_target(df, tgt, response_col = "response", x_limits = NULL, ...)
}
\arguments{
\item{df}{Dataframe containing data to plot. Must contain columns:
\itemize{
\item treatment
\item target
\item log_dose
\item response column, specified in response_col argument
}}

\item{tgt}{Name of target to plot}

\item{response_col}{Name of column containing response data}

\item{x_limits}{Limits for x axis of plot. Optional: if not provided, will be
calculated automatically.}

\item{...}{Extra parameters to pass to \code{base_dose_response()}}
}
\value{
A ggplot2 plot of the dose-response effect of the specified target on
all treatments
}
\description{
Plot dose-response effects of one target on all treatments
}
