% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_treatment}
\alias{plot_treatment}
\title{Plot dose-response effects of one treatment on all targets}
\usage{
plot_treatment(df, trt, response_col = "response", x_limits = NULL, ...)
}
\arguments{
\item{df}{Dataframe containing data to plot. Must contain columns:
\itemize{
\item treatment
\item target
\item log_dose
\item response column, specified in response_col argument
}}

\item{trt}{Name of treatment to plot}

\item{response_col}{Name of column containing response data}

\item{...}{Extra parameters to pass to \code{base_dose_response()}}
}
\value{
A ggplot2 plot of the dose-response effect of the specified treatment
on all targets
}
\description{
Plot dose-response effects of one treatment on all targets
}
