% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_treatment}
\alias{plot_treatment}
\title{Plot dose-response effects of one treatment on all targets}
\usage{
plot_treatment(
  df,
  trt,
  response_col = "response",
  color_map = NULL,
  rigid = FALSE,
  x_limits = NULL,
  legend_title = ggplot2::waiver(),
  legend_labels = ggplot2::waiver(),
  ...
)
}
\arguments{
\item{df}{Dataframe containing data to plot. Must contain columns:
\itemize{
\item treatment
\item target
\item log_dose
\item response column, specified in response_col argument
}}

\item{trt}{Name of treatment to plot}

\item{response_col}{Name of column containing response data}

\item{color_map}{Optional vector of color specifiers (named vector is
recommended) for manually specifying colors. If not provided,
\code{viridis_range()} and \code{viridis::scale_color_viridis()} will be used to
determine colors.}

\item{rigid}{Whether to set rigid value for low-dose asymptote. 100 if
decreasing, 0 if increasing.}

\item{x_limits}{Limits for x axis of plot. Optional: if not provided, will be
calculated automatically.}

\item{legend_title}{Optional title for the legend. If not provided, will be
determined by \code{ggplot2::waiver()} and will be the name of the aesthetic
used for the legend.}

\item{legend_labels}{Optional named vector to manually set legend labels. If
not provided, will be determined by \code{ggplot2::waiver()} and will be the
names of the groups in the data.}

\item{...}{Extra parameters to pass to \code{base_dose_response()}}
}
\value{
A ggplot2 plot of the dose-response effect of the specified treatment
on all targets
}
\description{
Plot dose-response effects of one treatment on all targets
}
