% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_plate_data}
\alias{preprocess_plate_data}
\title{Preprocess plate-based dose-response data}
\usage{
preprocess_plate_data(df, treatments, targets)
}
\arguments{
\item{df}{A dataframe containing raw plate data}

\item{treatments}{Optional vector of treatments to filter for}

\item{targets}{Optional vector of targets to filter for}
}
\value{
A dataframe containing processed data.
}
\description{
Preprocess plate-based dose-response data
}
