% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_summarize.R
\name{summarize_models}
\alias{summarize_models}
\title{Summarize model parameters for all combinations of treatments and targets}
\usage{
summarize_models(data, response_col = "response_norm")
}
\arguments{
\item{data}{A dataframe containing dose-response data for one or more
combinations of treatments and targets. The dataframe should contain the
following columns:
\itemize{
\item "response_norm" (default) or other activity column name from \code{response_col}
\item log_dose: dose of treatment in log(molar) units
}}

\item{response_col}{Name of the column containing response values.
Default is "response". Typical alternatives might include e.g.
"percent_inhibition", "viability", "growth", "activity".}
}
\value{
A summary dataframe containing one row for each combination of
treatment and target in \code{data} and containing the following columns:
\itemize{
\item "model": the actual model object
\item "IC50_nM": the dose at which the response is exactly 50.
\item "EC50_nM": the dose at which 50\% of the maximum response is achieved, whether
the response is positive or negative.
\item "EC75_nM": the dose at which 75\% of the maximum response is achieved, whether
the response is positive or negative.
\item "hill_slope": the \code{eta} parameter from \code{\link[drda:drda]{drda::drda()}}. The rate of growth of
the model, unsigned (always positive).
\item "low_dose_asymptote": the value of the function as dose -> -Inf.
\item "high_dose_asymptote": the value of the function as dose -> Inf.
}
}
\description{
Summarize model parameters for all combinations of treatments and targets
}
