% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{summarize_response}
\alias{summarize_response}
\title{Summarize replicate data for plotting}
\usage{
summarize_response(df, response_col = "response")
}
\arguments{
\item{df}{Dataframe containing the data to plot}

\item{response_col}{Name of the column containing response data to summarize}
}
\value{
A dataframe containing the summarized data
}
\description{
Summarize replicate data for plotting
}
